+++ apps/bot/src/domains/economy/errors.ts
@@ -1,0 +1,7 @@
+import { AppError } from '../../discord/errors.js';
+
+export class InsufficientFundsError extends AppError {
+  constructor() {
+    super('Fonds insuffisants.', 'warn');
+  }
+}
