+++ apps/bot/src/domains/economy/index.ts
@@ -1,2 +1,3 @@
 export { formatBerry } from './utils/format-berry.js';
-export { buy, sell, InsufficientFundsError } from './service.js';
+export { buy, sell } from './service.js';
+export { InsufficientFundsError } from './errors.js';
